Dev thanh

+++ BackEnd/tasks/ntpos-microservices/order-service/src/main/java/fitnlu/ntpos/orderservice/adapter/input/adapter/ChangeOrderEndpointAdapter.java
@@ -1,6 +1,8 @@
 package fitnlu.ntpos.orderservice.adapter.input.adapter;
 
+import fitnlu.ntpos.grpcproto.ProductRequest;
 import fitnlu.ntpos.grpcproto.UserResponse;
+import fitnlu.ntpos.orderservice.adapter.gRPCInput.ProductGrpcClientService;
 import fitnlu.ntpos.orderservice.adapter.gRPCInput.UserGrpcClientService;
 import fitnlu.ntpos.orderservice.adapter.input.dto.*;
 import fitnlu.ntpos.orderservice.adapter.input.mapper.OrderLineItemMapperInput;
@@ -31,6 +33,7 @@ public class ChangeOrderEndpointAdapter implements IChangeOrderEndpointPort {
     private final IDeleteAllOrderLineItemsFromOrderUseCase deleteAllOrderLineItemsFromOrderUseCase;
     private final KafkaTemplate<String, Notification> kafkaTemplate;
     private final UserGrpcClientService userGrpcClientService;
+    private final ProductGrpcClientService productGrpcClientService;
     @Override
     public OrderOutput createOrder(OrderInput orderInput) {
         System.out.println("start create order: 1"+orderInput.userID());
@@ -38,15 +41,22 @@ public class ChangeOrderEndpointAdapter implements IChangeOrderEndpointPort {
         System.out.println(userResponse.getHaveUser()+" - "+userResponse.getVerified());
         System.out.println("start create order: 2");
         if(userResponse.getHaveUser()&& userResponse.getVerified()){
-            OrderOutput orderOutput = OrderMapperInput.toDTO(createOrderUseCase.createOrder(OrderMapperInput.toDomain(orderInput)));
-            kafkaTemplate.send(orderInput.userID(), Notification.builder()
-                    .description("CREATE ORDER")
-                    .userID(orderOutput.getUserID())
-                    .status(orderOutput.getStatus())
-                            .timestamp(String.valueOf(System.currentTimeMillis()/1000))
-                    .build());
-            System.out.println("start create order: 3");
-            return orderOutput;
+           boolean check =  productGrpcClientService.updateQuantity(orderInput.orderLineItems().stream().map(orderLineItemInput -> ProductRequest.newBuilder()
+                    .setProductID(orderLineItemInput.productID())
+                    .setQuantity(orderLineItemInput.quantity())
+                    .setPrice(orderLineItemInput.price())
+                    .build()).toList());
+           if(check) {
+               OrderOutput orderOutput = OrderMapperInput.toDTO(createOrderUseCase.createOrder(OrderMapperInput.toDomain(orderInput)));
+               kafkaTemplate.send(orderInput.userID(), Notification.builder()
+                       .description("CREATE ORDER")
+                       .userID(orderOutput.getUserID())
+                       .status(orderOutput.getStatus())
+                       .timestamp(String.valueOf(System.currentTimeMillis() / 1000))
+                       .build());
+               return orderOutput;
+           }else
+               throw new RuntimeException("Product not found or not enough quantity");
         }
        else throw new RuntimeException("User not found or not verified");
 
@@ -63,8 +73,17 @@ public class ChangeOrderEndpointAdapter implements IChangeOrderEndpointPort {
         OrderOutput orderOutput = OrderMapperInput.toDTO(updateOrderUseCase.updateOrder(orderID, OrderMapperInput.toDomain(orderInput)));
         if(orderInput.orderLineItems()!= null) {
             System.out.println("Update order 02");
-            deleteAllOrderLineItemsFromOrderUseCase.deleteAllOrderLineItemsFromOrder(orderID);
-            addOrderLineItemFromOrder(orderID, orderInput.orderLineItems());
+            boolean check =  productGrpcClientService.updateQuantity(orderInput.orderLineItems().stream().map(orderLineItemInput -> ProductRequest.newBuilder()
+                    .setProductID(orderLineItemInput.productID())
+                    .setQuantity(orderLineItemInput.quantity())
+                    .setPrice(orderLineItemInput.price())
+                    .build()).toList());
+            if(check) {
+                deleteAllOrderLineItemsFromOrderUseCase.deleteAllOrderLineItemsFromOrder(orderID);
+                addOrderLineItemFromOrder(orderID, orderInput.orderLineItems());
+            }else{
+                throw new RuntimeException("Product not found or not enough quantity");
+            }
             System.out.println("Update order 03");
         }
         if(orderInput.tables()!= null) {
